bookmark: .gitlab-ci.yml: Fix command in before_script
The command in the before_script section of the eslint job was fixed to correctly generate the changelog.html file. Previously, the bash command was missing, causing the script to fail.

These changes were done to ensure that the changelog.html file is generated correctly during the CI/CD process.