✨ feat(gitlab-ci.yml): add git package for changelog creation
Add git package to the before_script section of the eslint job in .gitlab-ci.yml. This change ensures that the changelog creation process can use the git command.

The git command is required to fetch the commit information and generate the changelog.html file. By adding the git package, we provide the necessary dependency for running the changelog creation script.