:art: chore(compatibility): update docker tag page size
The page size of the docker tag API has been updated from 100 to 30
to improve performance and reduce load on the API.

:bug: fix(compatibility): fix query test assertion

The query test assertion has been fixed to check if the query data
is not an empty string instead of checking for an exact match with an object.

:bug: fix(compatibility): fix file test assertion

The file test assertion has been fixed to check if the loaded data
is not an empty string instead of checking for an exact match with an object.

:bug: fix(compatibility): fix server test assertions

The server test assertions have been fixed to check if the scan,
connect with hard-coded credentials, and guessed credentials data
is not an empty string instead of checking for an exact match with an object.

:recycle: refactor(compatibility): change MongoDB to MySQL

The compatibility scenarios have been updated to use MySQL instead