feat(compatibility): Update docker tags page size
This commit updates the page size for fetching docker tags from 30 to
150 in the compatibility helper.js file.

fix(front): Adjust background color in styles.scss

In the styles.scss file, the background color for the material-symbols-outlined
class selector is adjusted to rgba(30, 30, 30, 0.6) and a border of 1px solid
#1de9b6 is added.

style(front): Decrease font size in code and .hljs

The font size for the code and .hljs selectors is decreased to 12px in the
styles.scss file.

style(front): Add border to .actions selector in column.component.scss

A border of 1px solid #0f80cc is added to the .actions selector in the
column.component.scss file.

style(front): Remove font-size from logs-dialog.html

The font-size property is removed from the style attribute in the