:building_construction: chore(ci): update docker build command
Update the docker build command in the `.gitlab-ci.yml` file
to include the `--sbom=true` and `--provenance=true` flags.
Previously, the command only included the `--platform` and `-t` flags.

This commit ensures that the build process generates a Software Bill of Materials (SBoM)
and includes the provenance information for the image.

No additional changes were made in this commit.