:sparkles: fix(compatibility): Fix issue with running webdb in CI
The webdb container was not being properly launched in the CI environment due to a missing call to the `runWebDB` function. This function is responsible for running the necessary Docker commands to start the webdb container. By adding the call to `runWebDB` in the `app.js` file, the issue is resolved.

+ Fix issue with running webdb in CI