feat(front): add first-visit-dialog component
Add the first-visit-dialog component that is displayed when the user visits
the demo site for the first time. This component is added to the connection
component, and it is opened as a dialog using Material Dialog. The first-visit
dialog provides important information about the demo site limitations and
encourages the user to provide feedback.

Changes Made:
- Created the first-visit-dialog HTML and CSS files.
- Added the FirstVisitDialog component to the connection.component.ts file.
- Added a check in the connection.component.ts to display the first-visit-dialog
  when the user visits the demo site for the first time.