:bug: fix: fix vulnerable package dependencies
Fixes a vulnerability in the package dependencies by updating the "@faker-js/faker" and "@ngneat/falso" packages to their latest versions.

:recycle: refactor: remove unused code

Removes unused code in the database controller file.

:lipstick: style: update styles for connection component

Updates the styles for the connection component to improve the layout and responsiveness.

:sparkles: feat: add new feature to AI component

Adds a new feature to the AI component that allows users to save their chat history.

:hammer: chore: update Monaco Editor options

Updates the Monaco Editor options in the edit-connection component to customize the editor's behavior and appearance.