feat(driver): Add support for altering complex objects in MySQL and MariaDB
Add the ability to alter complex objects in MySQL and MariaDB drivers by implementing the `alterComplex` method in the respective driver classes. This method is used to generate the SQL statements necessary to alter complex objects such as triggers, functions, procedures, and checks.

The `alterComplex` method takes two parameters: the complex object to be altered (`Complex`) and the type of alteration (`string`). Depending on the type, the method generates the appropriate SQL statements.

This feature allows users to easily modify and update complex objects in their databases without having to manually write the SQL statements themselves. It enhances the flexibility and convenience of the MySQL and MariaDB drivers.

This commit also includes some code clean-up and minor bug fixes.