:art: chore: optimize SQL cleanQuery function
The cleanQuery function in the SQL class was optimized to remove comments, convert multi-line queries to single line, and trim unnecessary white spaces at the beginning and end of the query. Previously, the function also removed the semicolon at the end of the query, but this functionality was removed to ensure consistency. This commit improves the performance and readability of the code.