:hammer: fix(wrapper): Fix escaping issue on function and procedure names
The issue was caused by not properly escaping function and procedure names
in the MySQL wrapper. This commit fixes the issue by using the `escapeId`
function to escape the names when executing the `SHOW CREATE FUNCTION` and
`SHOW CREATE PROCEDURE` commands.

The commit also includes additional changes related to the `CHECK` constraint
in the MySQL and MariaDB wrappers. The `CHECK` constraint value is now properly
concatenated with the table name using the `CONCAT` function in both wrappers.

Additionally, in the MariaDB wrapper, the `NULL` value for `COLUMN_DEFAULT` is
now set to `null` instead of the string `"NULL"`.

This commit addresses these issues and improves the overall functionality and
reliability of the `mysql.js` wrapper file.