:rotating_light: fix(front): fix resizing issue in table headers
The table headers in the front/src/styles.scss were not resizing properly. This commit fixes the issue by removing the unnecessary resize-holder styles from the .mat-header-cell:not(:nth-last-of-type(-n+2)) .resize-holder selector.

:bug: fix(front): update validName regex in helper.ts

In the front/src/shared/helper.ts file, the validName regex was not allowing names with more than 400 characters. This commit updates the regex to allow names with 1 to 400 characters.

:recycle: refactor(front): refactor form creation in structure.component.ts

In the front/src/app/core/structure/structure.component.ts file, the form creation code was not using proper dependency injection. This commit refactors the code to use proper dependency injection for FormBuilder.

:art: style(front): adjust height of container in query.component.scss

In the front/src/app/core/query/query.component.scss file, the height